+++ test/case_test.exs
@@ -2,16 +2,24 @@ defmodule Tux.CaseTest do
   require Tux.Case
   Tux.Case.__using__()
 
-  program "program" do
-    command "foo", do: {:ok, "bar"}
-    command "bar", do: {:ok, "foo"}
-  end
+  describe "macros work & keep their formatting" do
+    program "geometry" do
+      command "shape", do: {:ok, "square"}
 
-  test "exact scenario" do
-    assert execute("program", invoke: "foo", expect: [exactly: "bar\n"])
-  end
+      dispatcher "more" do
+        command "shapes", do: {:ok, "square rectangle"}
+      end
+    end
 
-  test "approx scenario" do
-    assert execute("program", invoke: "bar", expect: [approx: "f"])
+    scenario "geometry shape",
+      using: "geometry",
+      invoke: "shape",
+      expect: [approx: "square"]
+
+    test "geometry more shapes" do
+      execute "geometry",
+        invoke: "more shapes",
+        expect: [exactly: "square rectangle\n"]
+    end
   end
 end